11.6 20:47 miao
2014.11.06
 1.瀑布流调整。
 2.新界面萌星推荐。

+++ Model/PhotoModel.h
@@ -31,6 +31,6 @@
 
 @property(nonatomic)int width;
 @property(nonatomic)int height;
-@property(nonatomic)int cmtWidth;
-@property(nonatomic)int cmtHeight;
+@property(nonatomic)float cmtWidth;
+@property(nonatomic)float cmtHeight;
 @end
